+++ Source/Objects/Hardware/USB/LakeShore336/ORLakeShore336Input.m
@@ -273,8 +273,8 @@ NSString* ORLakeShore336InputTemperatureChanged  = @"ORLakeShore336InputTemperat
 
 - (void) timeRateAtIndex:(int)i x:(double*)xValue y:(double*)yValue
 {
-    int count   = [timeRate count];
-    int index   = count-i-1;
+    NSUInteger count   = [timeRate count];
+    NSUInteger index   = count-i-1;
     *xValue     = [timeRate timeSampledAtIndex:index];
     *yValue     = [timeRate valueAtIndex:index];
 }
