+++ Source/Objects/Hardware/Vme/VMEcpu/ORVmecpuModel.h
@@ -79,43 +79,43 @@
 
 
 - (void) readLongBlock:(unsigned long *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberLongs
 		   withAddMod:(unsigned short) anAddressModifier
 		usingAddSpace:(unsigned short) anAddressSpace;
 
 - (void) writeLongBlock:(unsigned long *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberLongs
 			withAddMod:(unsigned short) anAddressModifier
 		 usingAddSpace:(unsigned short) anAddressSpace;
 
 - (void) readLong:(unsigned long *) readAddress
-	   atAddress:(unsigned int) vmeAddress
+	   atAddress:(unsigned long) vmeAddress
 	 timesToRead:(unsigned int) numberLongs
 	  withAddMod:(unsigned short) anAddressModifier
    usingAddSpace:(unsigned short) anAddressSpace;
 
 - (void) readByteBlock:(unsigned char *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberBytes
 		   withAddMod:(unsigned short) anAddressModifier
 		usingAddSpace:(unsigned short) anAddressSpace;
 
 - (void) writeByteBlock:(unsigned char *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberBytes
 			withAddMod:(unsigned short) anAddressModifier
 		 usingAddSpace:(unsigned short) anAddressSpace;
 
 - (void) readWordBlock:(unsigned short *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberWords
 		   withAddMod:(unsigned short) anAddressModifier
 		usingAddSpace:(unsigned short) anAddressSpace;
 
 - (void) writeWordBlock:(unsigned short *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberWords
 			withAddMod:(unsigned short) anAddressModifier
 		 usingAddSpace:(unsigned short) anAddressSpace;
